budgetthingy/flask3/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.2
frozenlist==1.4.1
google==3.0.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
nulltype==2.3.1
oauthlib==3.2.2
plaid-python==20.0.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
soupsieve==2.5
twilio==9.0.2
uritemplate==4.1.1
urllib3==2.2.1
werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.1
